+++ app/src/main/java/eu/kanade/tachiyomi/data/track/kitsu/Kitsu.kt
@@ -198,7 +198,7 @@ class Kitsu(id: Long) : BaseTracker(id, "Kitsu"), DeletableTracker {
 
     // RK --> autofill entry metadata (Fill from tracker)
     override suspend fun getMangaMetadata(track: DomainTrack): TrackMangaMetadata {
-        return api.getMangaMetadata(track)
+        return api.getMangaMetadata(track, includeAdult = trackPreferences.showAdultTrackerContent.get())
     }
     // RK <--